fixed cards not hiding due to depth dabbeling
earlier today i wrote the code to make the selected card always have a lesser depth then the other cards, as to make them show on top. little did i know that messing with depth makes it so that layer_set_visible does not work for the cards. the fix; loop over all card instances and deactivate them when pressing esc, (this will have to be changed later because if you find the combination then also the menu will close without pressing esc. and i reactivate all the instances in the room because only activating the card instances did not work.